Remove Release type condition for artifact publishing (Mac)
This commit removes the 'Release' `B_BUILD_TYPE` condition for the Mac
builder, and instead just keeps the condition of the image ONLY to be
built by the Big Sur build agent.

Now, the reason for that condition is because I am future proofing the
CI configuration for when we ship M1 builds. Currently, we do not. The
other reason, but also more pressing, is to avoid multiple artifacts
being published. There *is* a possibility the Big Sur build agent
becomes incompatible with older macOS releases, but the CMake
configuration would seem to suggest there's backwards compatibility
being available.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>